fix conflicts with upstream

+++ egs/TEMPLATE/svs1/pyscripts/audio/format_wav_midi_scp.py
@@ -30,7 +30,7 @@ def get_parser():
     parser.add_argument("outdir", type=str, help="segmented data folder path")
     parser.add_argument(
         "fs",
-        type=np.int16,
+        type=np.int32,
         default=None,
         help="If the sampling rate specified, " "Change the sampling rate.",
     )
@@ -77,8 +77,8 @@ def segmentation(
     out_mididir = Path(args_outdir) / "midi"
 
     if args_fs is not None and args.fs != rate:
-        wave = resampy.resample(wave.astype(np.float64), rate, args_fs, axis=0)
-        wave = wave.astype(np.int16)
+        wave = resampy.resample(wave, rate, args_fs, axis=0)
+        # wave = wave.astype(np.int16)
         rate = args_fs
 
     sub_wav = get_subsequence(segment_begin, segment_end, wave, rate)
@@ -138,14 +138,18 @@ if __name__ == "__main__":
 
     wav_reader = SoundScpReader(args.scp + "/wav.scp")
     if args.fs is not None:
-        midi_reader = MIDIScpReader(args.scp + "/midi.scp", rate=np.int16(args.fs))
+        midi_reader = MIDIScpReader(args.scp + "/midi.scp", rate=np.int32(args.fs))
     else:
         midi_reader = MIDIScpReader(args.scp + "/midi.scp")
 
     if args.fs is None:
         args.fs = midi_reader.rate
     # Note: generate segmented file
-    wav_writer = SoundScpWriter(out_wavdir, out_wavscp, format="wav",)
+    wav_writer = SoundScpWriter(
+        out_wavdir,
+        out_wavscp,
+        format="wav",
+    )
     midi_writer = MIDIScpWriter(out_mididir, out_midiscp, format="midi", rate=args.fs)
 
     for key, val in segments.items():
@@ -165,3 +169,5 @@ if __name__ == "__main__":
             note_seq,
             tempo_seq,
         )
+
+
